TPS_Cassifier_v1_KMeans/requirements.txt
pandas>=1.5.0
scikit-learn>=1.3.0
xgboost>=1.7.0
biopython>=1.81
h5py>=3.8.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
joblib>=1.3.0
tqdm>=4.65.0

